datatype conversion

+++ public-api/src/main/java/org/pmiops/workbench/publicapi/DataBrowserController.java
@@ -894,9 +894,9 @@ public class DataBrowserController implements DataBrowserApiDelegate {
         float binWidth = (max-min)/11;
 
         if (min >= 10 && max >= 10) {
-            binWidth = Math.ceil(binWidth/10)*10;
+            binWidth = (float)Math.ceil(binWidth/10)*10;
         } else if ((max-min) <= 1 && (max-min) >= 0.1) {
-            binWidth = Math.ceil(binWidth/0.1)*0.1;
+            binWidth = (float) Math.ceil(binWidth/0.1)*0.1;
         }
 
         bins.add(Float.valueOf(String.format("%.2f", min+binWidth)));
